@@ -1,55 +1,96 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
 
 namespace MyLab.ApiClient
 {
     /// <summary>
-    /// Web API client factory
+    /// Provides abilities to send requests to web API
     /// </summary>
+    /// <typeparam name="TContract">API contract</typeparam>
     public class ApiClient<TContract>
-        where TContract : class 
     {
-        /// <summary>
-        /// Api client factory
-        /// </summary>
-        public static readonly ApiClient<TContract> Factory = new ApiClient<TContract>();
+        private readonly ServiceDescription<TContract> _description;
+        private readonly IHttpClientProvider _httpClientProvider;
 
-        ApiClient()
+        private ApiClient(ServiceDescription<TContract> description, IHttpClientProvider httpClientProvider)
         {
+            _description = description ?? throw new ArgumentNullException(nameof(description));
+            _httpClientProvider = httpClientProvider ?? throw new ArgumentNullException(nameof(httpClientProvider));
+        }
+        
+        /// <summary>
+        /// Creates API client based on http client factory with specified service name
+        /// </summary>
+        public static ApiClient<TContract> Create(IHttpClientProvider httpClientProvider)
+        {
+            return new ApiClient<TContract>(
+                ServiceDescription<TContract>.Create(),
+                httpClientProvider);
+        }
+
+        public ApiRequest<string> Call(Expression<Action<TContract>> serviceCallExpr)
+        {
+            if(!(serviceCallExpr.Body is MethodCallExpression mExpr))
+                throw new NotSupportedException("Only method calls are supported");
             
+            var exprParams = GetParametersForExpression(mExpr);
+            
+            return new ApiRequest<string>(
+                _description.Url, 
+                exprParams.MethodDescription,
+                exprParams.Appliers,
+                _httpClientProvider);
         }
 
-        /// <summary>
-        /// Creates web api client 
-        /// </summary>
-        public TContract Create(IHttpClientProvider httpClientProvider, IHttpMessagesListener httpMessagesListener = null)
+        public ApiRequest<TRes> Call<TRes>(Expression<Func<TContract, TRes>> serviceCallExpr)
         {
-            var httpInvoker = new HttpRequestInvoker(httpClientProvider);
-
-            var clientStrategy = new WebClientProxyStrategy(httpInvoker)
-            {
-                HttpMessagesListener = httpMessagesListener
-            };
-            var description = ApiClientDescription.Get(typeof(TContract));
-
-            return ClientProxy<TContract>.CreateProxy(description, clientStrategy);
+            if(!(serviceCallExpr.Body is MethodCallExpression mExpr))
+                throw new NotSupportedException("Only method calls are supported");
+            
+            var exprParams = GetParametersForExpression(mExpr);
+            
+            return new ApiRequest<TRes>(
+                _description.Url, 
+                exprParams.MethodDescription,
+                exprParams.Appliers,
+                _httpClientProvider);
         }
 
-        /// <summary>
-        /// Creates web api client 
-        /// </summary>
-        public TContract Create(string baseUrl, TimeSpan? timeout = null, IDictionary<string, string> headers = null)
+        (MethodDescription MethodDescription, IEnumerable<IParameterApplier> Appliers) GetParametersForExpression(MethodCallExpression serviceCallExpr)
         {
-            var httpClientProvider = new DefaultHttpClientProvider(baseUrl)
-            {
-                Timeout = timeout.GetValueOrDefault(),
-                Headers = headers == null 
-                    ? null
-                    : new Dictionary<string, string>(headers)
+            if(!(serviceCallExpr is MethodCallExpression mExpr))
+                throw new NotSupportedException("Only method calls are supported");
+            if(!_description.Methods.TryGetValue(mExpr.Method.MetadataToken, out var mDesc))
+                throw new ApiClientException("Specified method description not found");
 
-            };
+            var args = mExpr.Arguments;
+            
+            if(args.Count != mDesc.Parameters.UrlParams.Count+ mDesc.Parameters.ContentParams.Count+ mDesc.Parameters.HeaderParams.Count)
+                throw new ApiClientException("ParamAppliers number mismatch");
+            
+            var callParams = new List<IParameterApplier>();
 
-            return Create(httpClientProvider);
+            callParams.AddRange(mDesc.Parameters.UrlParams.Select(d => 
+                new UrlParameterApplier(
+                    d, 
+                    new DefaultApiRequestParameterValueProvider(args[d.Position])
+                    )));
+
+            callParams.AddRange(mDesc.Parameters.HeaderParams.Select(d =>
+                new HeaderParameterApplier(
+                    d,
+                    new DefaultApiRequestParameterValueProvider(args[d.Position])
+                )));
+
+            callParams.AddRange(mDesc.Parameters.ContentParams.Select(d =>
+                new ContentParameterApplier(
+                    d,
+                    new DefaultApiRequestParameterValueProvider(args[d.Position])
+                )));
+
+            return (mDesc, callParams);
         }
     }
 }
